sign in page added and implemented into render. Personal info moved into private routes. Me request was also implemented into private routes
