+++ src/pages/sign-in-page/sign-in-page.tsx
@@ -1,0 +1,5 @@
+import { SignIn } from '@/components/auth/sign-in'
+
+export const SignInPage = () => {
+  return <SignIn onSubmit={() => {}} />
+}
